requirements.txt
# AI Image Detector GUI - Dependencies
# Install with: pip install -r requirements.txt

torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
customtkinter>=5.2.0
flask>=3.0.0
